requirements.txt
google-generativeai
Pillow
Flask
Flask-Cors
